implemented grpc

+++ distributed-systems/src/raft/messages/request_vote.go
@@ -3,8 +3,8 @@ package messages
 type RequestVoteRequest struct {
 	Term         uint64
 	CandidateId  uint32
-	LastLogIndex uint32
-	LastLogTerm  uint32
+	LastLogIndex int32
+	LastLogTerm  uint64
 }
 
 type RequestVoteResponse struct {
